requirements_dev.txt
-i https://pypi.org/simple
appdirs==1.4.3
atomicwrites==1.1.5
attrs==18.1.0
beautifulsoup4==4.6.0
black==18.6b4
bumpversion==0.5.3
certifi==2018.4.16
chardet==3.0.4
click==6.7
coverage==5.0a1
django-webtest==1.9.3
docutils==0.14
factory-boy==2.11.1
faker==0.8.16
flake8==3.5.0
flit==1.0
freezegun==0.3.10
idna==2.7
inflection==0.3.1
mccabe==0.6.1
more-itertools==4.2.0
pluggy==0.6.0
py==1.5.4
pycodestyle==2.3.1
pyflakes==1.6.0
pytest-cov==2.5.1
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-django==3.3.2
pytest-factoryboy==2.0.1
pytest-pythonpath==0.7.2
pytest-sugar==0.9.1
pytest==3.6.2
python-dateutil==2.7.3
pytoml==0.1.16
requests-download==0.1.2
requests==2.19.1
six==1.11.0
termcolor==1.1.0
text-unidecode==1.2
toml==0.9.4
urllib3==1.23
waitress==1.1.0
webob==1.8.2; python_version != '3.1.*' and python_version != '3.2.*' and python_version != '3.0.*' and python_version >= '2.7'
webtest==2.0.30
